ocr_excel_app/requirements.txt
Flask
pillow
pytesseract
openpyxl